Changed uint64 -> uint32 for use on ARM32
See golang issue #599

+++ neighborhood.go
@@ -31,7 +31,7 @@ func (n *neighborhoodSet) insertNode(node Node, proximity int64) (*Node, error) 
 	return n.insertValues(node.ID, node.LocalIP, node.GlobalIP, node.Region, node.Port, node.routingTableVersion, node.leafsetVersion, node.neighborhoodSetVersion, proximity)
 }
 
-func (n *neighborhoodSet) insertValues(id NodeID, localIP, globalIP, region string, port int, rTVersion, lSVersion, nSVersion uint64, proximity int64) (*Node, error) {
+func (n *neighborhoodSet) insertValues(id NodeID, localIP, globalIP, region string, port int, rTVersion, lSVersion, nSVersion uint32, proximity int64) (*Node, error) {
 	n.lock.Lock()
 	defer n.lock.Unlock()
 	if id.Equals(n.self.ID) {
